Widget files created to shorten the code
